--- module-application/src/main/java/com/wsws/moduleapplication/feed/dto/AnswerFindServiceResponse.java
@@ -1,5 +1,0 @@
-package com.wsws.moduleapplication.feed.dto;
-
-
-public record AnswerFindServiceResponse() {
-}
