+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/AnswerCreateServiceResponse.java
@@ -1,4 +1,4 @@
-package com.wsws.moduleapplication.feed.dto;
+package com.wsws.moduleapplication.feed.dto.answer;
 
 public record AnswerCreateServiceResponse(
         Long answerId
