+++ module-application/src/main/java/com/wsws/moduleapplication/feed/dto/answer/AnswerFindServiceResponse.java
@@ -1,0 +1,5 @@
+package com.wsws.moduleapplication.feed.dto.answer;
+
+
+public record AnswerFindServiceResponse() {
+}
